requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
click==8.3.1
Django==5.2.8
django-cors-headers==4.9.0
django-cuser==2017.3.16
django-modeltranslation==0.19.17
django-tinymce==5.0.0
frozenlist==1.8.0
h11==0.16.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
pillow==12.0.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.2.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
yarl==1.22.0
